Port build config to meson
This change introduces a new Meson configuration and removes the CMake
one.

Meson can manage dependencies _and_ run the tests, and is a convenient
and popular cross platform choice for C and C++ build setups nowadays.

We also dropped the need for Conan dependency manager (as Meson takes
care of that).

We also decided to migrate the test suit from CUnit to Unity for these
main reasons:

- Unity seems to be more up to date and maintained (than CUnit)
  nowadays.
- Unity has a similar API to CUnit so the tests didn't require too
  many changes.
- Unity has a more straightforward setup.
- Unity can be managed through `meson wrap`, thus not requiring manual
  installation or another dependency manager (like Conan).